requirements.txt
Django==2.1.1
djangorestframework==3.8.2
psycopg2==2.7.5
drf-extensions==0.4.0
django-rest-swagger==2.2.0